refactor(prover-client): extract ProvingScheduler base class for orchestrator infra
`ProvingOrchestrator` and `TopTreeOrchestrator` had duplicated copies of the same
broker-job submission envelope (~80 lines apiece): the `pendingProvingJobs`
controller list, the `SerialQueue` lifecycle, the `cancel`/`stop` plumbing, and
the `deferredProving<T>(state, request, callback)` wrapper that drops obsolete
results and routes errors to `state.reject(...)`.

Lift these to a new abstract `ProvingScheduler` base class. The minimal state
contract is `ProvingStateLike { verifyState(): boolean; reject(reason: string):
void }`, which both `EpochProvingState` / `CheckpointProvingState` /
`BlockProvingState` and `TopTreeProvingState` satisfy. The base owns:

  - `pendingProvingJobs`, `deferredJobQueue`, `getNumPendingProvingJobs`
  - `resetSchedulerState(abortJobs)` — drain + recreate queue, optionally abort
    in-flight jobs (the per-call abort flag covers both parent's
    `cancelJobsOnStop` config and top-tree's `{abortJobs}` arg)
  - `stop()` — standard "grab old queue, cancelInternal, await drain"
  - `deferredProving<S, T>(state, request, callback, isCancelled?)` — unified
    submit envelope. The `isCancelled` predicate covers top-tree's `cancelled`
    flag; the parent uses the default `() => false` and relies on `verifyState`.

Subclasses define `cancelInternal()` for their own cleanup (closing world-state
forks for the parent, propagating cancel into the proving state for top-tree).

Net code reduction: ~120 lines across the two orchestrators. The merge / padding
/ root rollup methods stay subclass-specific — they depend on state-class
methods that aren't unified here.
